fix reviewer comments
